Simplify debouncing code
Use the time.AfterFunc function which is a bit more straightforward.
This also solves a problem where the timeout code would execute when the
program launched, caused by the fact that a timer immediately starts
when it is created. This would cause the output view to be updated
before any input was entered, which resulted in incorrect output when a
custom filter string was entered.