Use a Buffer object instead of a Scanner
The bufio.Scanner reads one line at a time and has a maximum default
scan size of 64 KiB. JSON is often minified and stuffed onto a single
line, which causes the Scanner to choke.

Instead, use a simple bytes.Buffer object and just read stdin directly
into that. The buffer automatically resizes itself as necessary and is
perfectly capable of handing large JSON documents on a single line.