Remove blank entry from history
strings.Split() includes an empty string as the last element (due to the
trailing newline character) which creates an empty entry in the history
completion. We could also slice the last element off of the array (after
first checking to make sure that the array has non-zero length), but
using a bufio.Scanner is more robust.