Ignore some options until final output
Some options like -c (compact) or -M (monochrome) shouldn't be used when
filtering the input for previewing in ijq, but they should be used when
generating the final output when ijq exits. This means that some options
should be ignored until the app exits.

To do this, we can check whether or not we are writing to a TextView and
adjust the options accordingly. Further, we want to mimic jq's behavior
and default to using colored output when writing to a tty but using
non-colored output when writing to a file or pipe, so when the app exits
we check the status of stdout to see whether or not its a terminal. We
also provide the -C flag to *always* write colored output, even when
writing to a pipe or file.