Customize autocomplete list colors
In order to do this I have to use a fork of the tview library that
exposes functions to set the autocomplete list colors. The upstream
library hardcodes these styles in a way that it's essentially impossible
to get something that looks good on both light and dark terminal
backgrounds. If/when upstream adds this functionality I will revert the
dependency back.