Add more keybindings
Wrapping is now disabled and the Left/Right arrow keys (or h/l) can be
used to scroll the viewing panes horizontally.

Add support for "readline" bindings such as C-v and M-v to scroll the
viewing panes vertically, and C-a and C-e to scroll horizontally.

Also use 'u', 'f', 'b', and 'd' to scroll like in a pager.