Do not recalculate line count in scroll indicator update
Recalculating the line count can be expensive and does not need to be
recalculated everytime the scroll indicator is updated. In particular,
the input view has a fixed line count and so the line count can be
calculated when the input is first read and parsed.

The output view's line count must be updated every time the output text
changes.

We do not use tview's GetOriginalLineCount() function as it is quite
expensive (it iterates through all of the grapheme clusters in the text
and looks for line breaks). Instead, we use the basic method of "count
the number of newlines" in the text. I will wait to see if there is any
scenario for which this method is insufficient, but until then it's good
enough (and much faster).