Process new documents via a channel
Every time the filter input changes, send a new Document object over the
channel. A separate goroutine waits on the channel and runs the jq
process in a separate thread. This keeps the main UI thread completely
unblocked.

This is still not working quite right. When a new document is received
on the channel, we drop any other documents received until the current
jq process finishes. For long running processes, this means we miss
events and if a user types in a filter while a jq process is running
they will never see those updates apply.

What I need to do is use only the most recently received document on the
channel, but I can't figure out how to do this in Go. Channels act like
FIFOs so every item placed in the channel must be used, and there's no
way (that I can find) to have a single element channel that lets you
overwrite whatever is currently in there (like a mailbox).