Use a condition variable and update doc in place
This works better than a channel since it doesn't implement FIFO
behavior. Whenever the filter input is changed, we claim the lock and
update `doc` with the new filter, set `pending` to true, and signal the
condition variable. The separate goroutine waits for `pending` to be set
to true and makes a copy of whatever the most recent document is before
running jq in a subprocess. This ensures that we always run jq with the
most recent filter.