Override tview escape pattern for style tags
tview escapes strings of the form `[0]' (converting them to `[0[]')
because it interprets them as style tags. However, this is not a valid
style tag (as far as I can tell) so the escaped tag is not subsequently
stripped by tview, so the escaped version is what ends up being used by
the autocompleter.

Fixes: https://todo.sr.ht/~gpanders/ijq/11